Now first explains what the content of the docs is
I fixed the first part of the introduction so that it first explains what the starter plugin documentation is for before explaining I wrote it. I also added mentions that there will be examples of hooks, filters, views, and controllers in the last section when defining what each is in the earlier sections.